+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changes
 
-- Fixes to zombified piglins
-- Fix crash bugs with aqueducts
-- Make soul clay more common
+- Added Ancient Altars
+- Added Lost Pages
+- Added a sacrifice system, utilizing Lost Pages (found in chests) and Ancient Altars (crafted) to transform items into other items, at a cost. More recipes for this are TBD.
+- Added Slimed Netherrack, an ore.
+- Added a config for whether the portal-opening sacrifice works or not.
+- Added some recipes for making Blackstone from Soot and overworld rocks.